move modular controls into sub-namespaces
