+++ Content.Client/_MORBIT/Lobby/UI/CharacterEditor/AppearanceBodyOptions.xaml.cs
@@ -8,7 +8,7 @@ using Robust.Client.UserInterface.XAML;
 using Robust.Shared.Prototypes;
 
 #pragma warning disable IDE1006
-namespace Content.Client._MORBIT.Lobby.UI;
+namespace Content.Client._MORBIT.Lobby.UI.CharacterEditor;
 #pragma warning restore IDE1006
 
 [GenerateTypedNameReferences]
