+++ Content.Client/_MORBIT/Lobby/UI/CharacterEditor/CharacterPreview.xaml
