+++ Content.Client/_MORBIT/Lobby/UI/CharacterEditor/CharacterPreview.xaml.cs
@@ -9,7 +9,7 @@ using Robust.Client.Utility;
 using Robust.Shared.Prototypes;
 
 #pragma warning disable IDE1006
-namespace Content.Client._MORBIT.Lobby.UI;
+namespace Content.Client._MORBIT.Lobby.UI.CharacterEditor;
 #pragma warning restore IDE1006
 
 [GenerateTypedNameReferences]
