+++ Content.Client/_MORBIT/Lobby/UI/CharacterEditor/ProfileButtons.xaml.cs
@@ -10,7 +10,7 @@ using Robust.Client.UserInterface.XAML;
 using Robust.Shared.ContentPack;
 
 #pragma warning disable IDE1006
-namespace Content.Client._MORBIT.Lobby.UI;
+namespace Content.Client._MORBIT.Lobby.UI.CharacterEditor;
 #pragma warning restore IDE1006
 
 [GenerateTypedNameReferences]
