+++ Content.Client/_MORBIT/Lobby/UI/CharacterEditor/Tabs/ProfileEditorAppearanceTab.xaml
@@ -1,7 +1,7 @@
 <BoxContainer
     xmlns="https://spacestation14.io"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:ui="clr-namespace:Content.Client._MORBIT.Lobby.UI"
+    xmlns:ui="clr-namespace:Content.Client._MORBIT.Lobby.UI.CharacterEditor"
     Orientation="Vertical">
     <ScrollContainer VerticalExpand="True">
         <BoxContainer Orientation="Vertical">
